cs467-capstone-main/requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.6
requests==2.23.0
haversine==2.5.1
google-cloud-storage==1.43.0
Pillow==8.4.0
werkzeug==1.0.1
flask-cors